requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
Django==1.10.4
djangorestframework==3.5.3
kombu==4.0.2
pytz==2016.10
redis==2.10.5
sarge==0.1.4
vine==1.1.3
